Extract everything we can and refactor
