+++ packages/edit-site/src/components/left-sidebar/navigation-panel/template-preview.js
